requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
sentence-transformers>=2.3.0
open-clip-torch>=2.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.66.0
gradio>=4.20.0
matplotlib>=3.7.0
seaborn>=0.13.0
huggingface-hub>=0.20.0
datasets>=2.16.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
scipy>=1.11.0
